security_project/requirements.txt
IP2Location>=8.10.5

Django>=3.2,<4.0
python-decouple>=3.6
psycopg2-binary>=2.9  


django-ipware>=4.0.0 
django-ipgeolocation>=1.0.0  

django-ratelimit>=3.0.0

celery>=5.2.0
redis>=4.0.0  


drf-yasg>=1.20.0 

gunicorn>=20.1.0  
whitenoise>=6.0.0  


django-cors-headers>=3.13.0


scikit-learn>=1.3.0
